adjusted the test
